tests/requirements-dev.txt
pytest
pytest-asyncio
pytest-cov
pytest-homeassistant-custom-component